requirements.txt

django-ckeditor==5.8.0
django-js-asset==1.2.2
Pillow==7.0.0
pytz==2019.3
